+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Handling nickname changes (first experimental version). New room_nickname_changed event.
 * Exporting NickNameChangeHandler.
 * New room_roster_initialized event.
+* Fix room_joined event.
 
 ## 0.2.10